+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: nleempoe <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/23 16:59:18 by nleempoe          #+#    #+#             */
-/*   Updated: 2021/04/23 17:03:00 by nleempoe         ###   ########.fr       */
+/*   Updated: 2021/04/27 12:34:55 by nleempoe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int		c;
+	size_t	c;
 	char	*s;
 	char	*d;
